requirements.txt
python-barcode==0.15.1
pillow>=9.0.0
